added [worskpaceId] to main dir - this changes every file in the dir
