+++ app/[workspaceId]/settings/members/add/page.tsx
@@ -44,7 +44,7 @@ export default function AddMemberPage() {
   return (
     <div className="flex flex-col gap-[28px]">
       <div className="flex flex-col gap-4">
-        <Link href="/settings/members">
+        <Link href="../members">
           <Button
             variant="link"
             className="text-tertiary flex items-center gap-2 p-2"
